requirements.txt
python-dotenv>=0.19.0
openai>=0.27.0
prompt_toolkit>=3.0.28
rich>=10.12.0
networkx>=2.6.3
pydantic>=1.9.0
pytest>=7.0.0
python-decouple>=3.6
aiohttp>=3.8.1
colorama>=0.4.4
typing-extensions>=4.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
